blockchain tests through TestSuite class
