+++ test/tools/libtesteth/TestSuite.cpp
@@ -33,12 +33,12 @@ namespace
 	// Return full path to folder for tests from _testFolder
 	fs::path getFullPathFiller(string const& _suiteFolder, string const& _testFolder)
 	{
-		return fs::path(test::getTestPath()) / "src" / fs::path(_suiteFolder + "Filler") / _testFolder;
+		return fs::path(test::getTestPath()) / "src" / fs::path(_suiteFolder) / _testFolder;
 	}
 
 	fs::path getFullPathTest(string const& _suiteFolder, string const& _testFolder)
 	{
-		return fs::path(test::getTestPath()) / "src" / _suiteFolder / _testFolder;
+		return fs::path(test::getTestPath()) / _suiteFolder / _testFolder;
 	}
 }
 
@@ -50,13 +50,13 @@ namespace test
 void TestSuite::runAllTestsInFolder(string const& _testFolder) const
 {
 	string const filter = test::Options::get().singleTestName.empty() ? string() : test::Options::get().singleTestName + "Filler";
-	std::vector<boost::filesystem::path> const files = test::getJsonFiles(getFullPathFiller(suiteFolder(), _testFolder).string(), filter);
+	std::vector<boost::filesystem::path> const files = test::getJsonFiles(getFullPathFiller(suiteFillerFolder(), _testFolder).string(), filter);
 	size_t fileCount = files.size();
 	if (test::Options::get().filltests)
 		fileCount *= 2; //tests are checked when filled and after they been filled
 
 	fs::path const destTestFolder = fs::path(suiteFolder()) / _testFolder;
-	fs::path const srcTestFolder = fs::path(suiteFolder() + "Filler") / _testFolder;
+	fs::path const srcTestFolder = fs::path(suiteFillerFolder()) / _testFolder;
 
 	auto suiteTestDo = [this](json_spirit::mValue const& _input, bool _fillin)
 	{
@@ -73,17 +73,18 @@ void TestSuite::runAllTestsInFolder(string const& _testFolder) const
 
 void TestSuite::copyAllTestsFromFolder(string const& _testFolder) const
 {
-	std::vector<fs::path> const files = test::getJsonFiles(getFullPathFiller(suiteFolder(), _testFolder).string());
+	std::vector<fs::path> const files = test::getJsonFiles(getFullPathFiller(suiteFillerFolder(), _testFolder).string());
 	for (auto const& file: files)
 	{
 		fs::path const destFile = getFullPathTest(suiteFolder(), _testFolder) / file.filename().string();
-		fs::path const srcFile = getFullPathFiller(suiteFolder(), _testFolder) / file.filename().string();
+		fs::path const srcFile = getFullPathFiller(suiteFillerFolder(), _testFolder) / file.filename().string();
 		clog << "Copying " << srcFile.string() << "\n";
 		clog << " TO " << destFile.string() << "\n";
 		auto testOutput = dev::test::TestOutputHelper();
 		dev::test::copyFile(srcFile.string(), destFile.string());
 		BOOST_REQUIRE_MESSAGE(boost::filesystem::exists(destFile.string()), "Error when copying the test file!");
 	}
+	runAllTestsInFolder(_testFolder); //check that copied tests are valid
 }
 
 }
